custom-scan: Stop PGrnExecCustomScan(), if PGroonga is already finalized
This commit is part of the work to implement a custom scan.

PGrnExecCustomScan() accesses a Groonga resource and returns the results one record per record. (Not implemented yet.)
If PGroonga is already finalized, `NULL` is returned because accessing Groonga resources when PGroonga is already finalized will cause a crash.
If PGrnExecCustomScan returns `NULL`, the scan ends there.